Renamed and moved some classes for ease of understanding location in UI
