+++ app/src/main/java/com/brainwallet/presenter/activities/BreadActivity.java
@@ -12,7 +12,6 @@ import android.net.Uri;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Handler;
-import android.telephony.TelephonyManager;
 import android.view.ViewTreeObserver;
 import android.view.animation.AnimationUtils;
 import android.widget.Button;
@@ -25,7 +24,6 @@ import androidx.activity.result.ActivityResultLauncher;
 import androidx.activity.result.contract.ActivityResultContracts;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
-import androidx.browser.customtabs.CustomTabsIntent;
 import androidx.constraintlayout.widget.ConstraintLayout;
 import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.core.app.ActivityCompat;
@@ -41,8 +39,6 @@ import com.brainwallet.navigation.Route;
 import com.brainwallet.presenter.activities.settings.SyncBlockchainActivity;
 import com.brainwallet.presenter.activities.util.BRActivity;
 import com.brainwallet.presenter.customviews.BRNotificationBar;
-import com.brainwallet.presenter.fragments.BuyTabFragment;
-import com.brainwallet.presenter.fragments.FragmentMoonpay;
 import com.brainwallet.presenter.history.HistoryFragment;
 import com.brainwallet.tools.animation.BRAnimator;
 import com.brainwallet.tools.animation.TextSizeTransition;
@@ -60,10 +56,9 @@ import com.brainwallet.tools.util.BRExchange;
 import com.brainwallet.tools.util.ExtensionKt;
 import com.brainwallet.tools.util.Utils;
 import com.brainwallet.ui.BrainwalletActivity;
-import com.brainwallet.ui.screens.home.SettingsViewModel;
+import com.brainwallet.ui.screens.settings.SettingsViewModel;
 import com.brainwallet.ui.screens.home.composable.HomeSettingDrawerComposeView;
 import com.brainwallet.ui.screens.home.receive.ReceiveDialogFragment;
-import com.brainwallet.ui.screens.home.receive.ReceiveDialogKt;
 import com.brainwallet.util.PermissionUtil;
 import com.brainwallet.wallet.BRPeerManager;
 import com.brainwallet.wallet.BRWalletManager;
