+++ app/src/main/java/com/brainwallet/ui/BentoActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.foundation.isSystemInDarkTheme
-import com.brainwallet.ui.screens.home.BentoMainScreen
+import com.brainwallet.ui.screens.home.MainScreen
 import com.grunt.brainwallet.core.presentation.theme.BrainwalletTheme
 
 class BentoActivity : ComponentActivity() {
@@ -15,7 +15,7 @@ class BentoActivity : ComponentActivity() {
         super.onCreate(savedInstanceState)
         setContent {
             BrainwalletTheme(isSystemInDarkTheme()) {
-                BentoMainScreen()
+                MainScreen()
             }
         }
     }
